requirements.txt
PyQt6==6.10.2
requests==2.32.3
pywin32; platform_system == "Windows"
pyobjc; platform_system == "Darwin"
vobject; platform_system == "Linux"